Resolve comment for mag
