+++ scaler/io/ymq/pymod_ymq/python.h
@@ -4,6 +4,8 @@
 #include <Python.h>
 #include <structmember.h>
 
+#include "scaler/io/ymq/pymod_ymq/utils.h"
+
 #if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
 static inline PyObject* Py_NewRef(PyObject* obj)
 {
@@ -39,16 +41,6 @@ static inline PyObject* PyObject_CallNoArgs(PyObject* callable)
     return PyObject_Call(callable, PyTuple_New(0), nullptr);
 }
 
-static inline PyObject* PyObject_CallOneArg(PyObject* callable, PyObject* arg)
-{
-    PyObject* args = PyTuple_Pack(1, arg);
-    if (!args)
-        return nullptr;
-    PyObject* result = PyObject_Call(callable, args, nullptr);
-    Py_DECREF(args);
-    return result;
-}
-
 static inline int PyModule_AddObjectRef(PyObject* mod, const char* name, PyObject* value)
 {
     Py_INCREF(value);  // Since PyModule_AddObject steals a ref, we balance it
@@ -60,10 +52,12 @@ static inline PyObject* PyType_FromModuleAndSpec(PyObject* pymodule, PyType_Spec
     (void)pymodule;
     if (!bases) {
         bases = PyTuple_Pack(1, (PyObject*)&PyBaseObject_Type);
-        if (!bases)
-            return nullptr;
+        if (bases) {
+            PyObject* res = PyType_FromSpecWithBases(spec, bases);
+            Py_DECREF(bases);  // avoid leak
+            return res;
+        }
         PyObject* res = PyType_FromSpecWithBases(spec, bases);
-        Py_DECREF(bases);  // avoid leak
         return res;
     }
     return PyType_FromSpecWithBases(spec, bases);
@@ -71,4 +65,91 @@ static inline PyObject* PyType_FromModuleAndSpec(PyObject* pymodule, PyType_Spec
 
 #define Py_TPFLAGS_IMMUTABLETYPE          (0)
 #define Py_TPFLAGS_DISALLOW_INSTANTIATION (0)
+#endif
+
+// NOTE: We define this no matter what version of Python we use.
+// an owned handle to a PyObject with automatic reference counting via RAII
+template <typename T = PyObject>
+class OwnedPyObject {
+public:
+    OwnedPyObject(): _ptr(nullptr) {}
+
+    // steals a reference
+    OwnedPyObject(T* ptr): _ptr(ptr) {}
+
+    OwnedPyObject(const OwnedPyObject& other) { this->_ptr = Py_XNewRef(other._ptr); }
+    OwnedPyObject(OwnedPyObject&& other) noexcept: _ptr(other._ptr) { other._ptr = nullptr; }
+    OwnedPyObject& operator=(const OwnedPyObject& other)
+    {
+        if (this == &other)
+            return *this;
+
+        this->free();
+        this->_ptr = Py_XNewRef(other._ptr);
+        return *this;
+    }
+    OwnedPyObject& operator=(OwnedPyObject&& other) noexcept
+    {
+        if (this == &other)
+            return *this;
+
+        this->free();
+        this->_ptr = other._ptr;
+        other._ptr = nullptr;
+        return *this;
+    }
+
+    ~OwnedPyObject() { this->free(); }
+
+    // creates a new OwnedPyObject from a borrowed reference
+    static OwnedPyObject fromBorrowed(T* ptr) { return OwnedPyObject((T*)Py_XNewRef((PyObject*)ptr)); }
+
+    // convenience method for creating an OwnedPyObject that holds Py_None
+    static OwnedPyObject none() { return OwnedPyObject((T*)Py_NewRef(Py_None)); }
+
+    bool is_none() const { return (PyObject*)_ptr == Py_None; }
+
+    // takes the pointer out of the OwnedPyObject
+    // without decrementing the reference count
+    // use this to transfer ownership to C code
+    T* take()
+    {
+        T* ptr     = this->_ptr;
+        this->_ptr = nullptr;
+        return ptr;
+    }
+
+    void forget() { this->_ptr = nullptr; }
+
+    // operator T*() const { return _ptr; }
+    explicit operator bool() const { return _ptr != nullptr; }
+    bool operator!() const { return _ptr == nullptr; }
+
+    T* operator->() const { return _ptr; }
+    T* operator*() const { return _ptr; }
+
+private:
+    T* _ptr;
+
+    void free()
+    {
+        if (!_ptr)
+            return;
+
+        if (!PyGILState_Check())
+            return;
+
+        Py_CLEAR(_ptr);
+    }
+};
+
+#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8
+static inline PyObject* PyObject_CallOneArg(PyObject* callable, PyObject* arg)
+{
+    OwnedPyObject<> args = PyTuple_Pack(1, arg);
+    if (!args)
+        return nullptr;
+    PyObject* result = PyObject_Call(callable, *args, nullptr);
+    return result;
+}
 #endif
